Breaks out methods from api.py to utils
